[Back] Move news in osmose website app
