+++ backend/api/models/__init__.py
@@ -24,6 +24,4 @@ from backend.api.models.annotations import (
     AnnotationTask,
 )
 
-from backend.api.models.news import News
-
 User = get_user_model()
