+++ backend/api/views/__init__.py
@@ -8,5 +8,4 @@ from backend.api.views.annotation_set import AnnotationSetViewSet
 from backend.api.views.annotation_campaign import AnnotationCampaignViewSet
 from backend.api.views.annotation_comment import AnnotationCommentViewSet
 from backend.api.views.annotation_task import AnnotationTaskViewSet
-from backend.api.views.news import NewsViewSet
 from backend.api.views.confidence_indicators import ConfidenceIndicatorSetViewSet
