+++ backend/osmosewebsite/admin/news.py
@@ -1,0 +1,8 @@
+from django.contrib.admin import ModelAdmin
+
+
+class NewsAdmin(ModelAdmin):
+    """News presentation in DjangoAdmin"""
+
+    list_display = ("title", "intro", "date", "vignette")
+    search_fields = ["title"]
