+++ backend/osmosewebsite/models/__init__.py
@@ -1,3 +1,4 @@
 """All Django models available"""
 
 from .team_member import TeamMember
+from .news import News
