+++ backend/osmosewebsite/serializers/__init__.py
@@ -1,2 +1,3 @@
 """OSmOSE Website API Serializers"""
 from .team_member import TeamMemberSerializer
+from .news import NewsSerializer
