+++ backend/osmosewebsite/serializers/news.py
@@ -1,11 +1,8 @@
 """Annotation task DRF serializers file"""
 
-# Serializers have too many false-positives on the following warnings:
-# pylint: disable=missing-function-docstring, abstract-method
-
 from rest_framework import serializers
 
-from backend.api.models import News
+from backend.osmosewebsite.models import News
 
 
 class NewsSerializer(serializers.ModelSerializer):
