+++ backend/osmosewebsite/views/__init__.py
@@ -1,2 +1,3 @@
 """OSmOSE Website API Views"""
 from .team_member import TeamMemberViewSet
+from .news import NewsViewSet
